Move validation check from cycle to face level
This validation check requires access to the surface, which I am in the
process of removing from `Cycle`.
